# documentation.docx/swagger.yaml
openapi: "3.0.1"
info:
  title: "Reviewer Module 2"
  description: "This repository provides the implementation of the primary module for the NITCONF conferencing website, specifically the Reviewer Module. This module serves as the primary interface for users who agree to analyze, research, and submit reviews for assigned research papers."
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
  version: '1.2'
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/AvinashRavanam/NITCONF/1.2
  - url: "http://localhost:8080"
    description: "Generated server url"
paths:
  /api/review:
    get:
      tags:
        - "Paper Review"
      summary: "Search for review"
      description: "Find the requested review and display it."
      operationId: "searchReview"
      parameters:
        - name: "action"
          in: "query"
          description: "The action to perform. Valid actions are 'view' or 'clear'."
          required: true
          schema:
            type: "string"
          example: "view"
        - name: "id"
          in: "query"
          description: "The ID of the paper."
          required: true
          schema:
            type: "integer"
            format: "int64"
          example: 1
        - name: "username"
          in: "query"
          description: "The username of the reviewer."
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    post:
      tags:
        - "Paper Review"
      summary: "Save Review"
      description: "Save the current progress in review."
      operationId: "saveReview"
      parameters:
        - name: "action"
          in: "query"
          description: "The action to perform. Valid actions are 'update' or 'save'."
          required: true
          schema:
            type: "string"
          example: "update"
        - name: "id"
          in: "query"
          description: "The ID of the paper."
          required: true
          schema:
            type: "integer"
            format: "int64"
          example: 1
        - name: "username"
          in: "query"
          description: "The username of the reviewer."
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewForm"
        required: true
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  schemas:
    ReviewForm:
      type: "object"
      properties:
        technical:
          type: "integer"
          format: "int32"
        significance:
          type: "integer"
          format: "int32"
        originality:
          type: "integer"
          format: "int32"
        appropriate:
          type: "integer"
          format: "int32"
        organization:
          type: "integer"
          format: "int32"
        style:
          type: "integer"
          format: "int32"
        refeersConfidence:
          type: "integer"
          format: "int32"
        overall:
          type: "integer"
          format: "int32"
        comment:
          type: "string"
    Review:
      type: "object"
      properties:
        id:
          $ref: "#/components/schemas/ReviewKey"
        technical:
          type: "integer"
          format: "int32"
        significance:
          type: "integer"
          format: "int32"
        originality:
          type: "integer"
          format: "int32"
        appropriate:
          type: "integer"
          format: "int32"
        organization:
          type: "integer"
          format: "int32"
        style:
          type: "integer"
          format: "int32"
        refeersConfidence:
          type: "integer"
          format: "int32"
        overall:
          type: "integer"
          format: "int32"
        comment:
          type: "string"
        reviewerStatus:
          type: "string"
        reviewStatus:
          type: "string"
    ReviewKey:
      type: "object"
...